pyproject.toml
[project]
name = "rtsp2jpeg-http-get"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python-headless",
    "python-dotenv",
    "docopt-ng"
]

[project.scripts]
rtsp2jpeg = "rtsp2jpeg_http_get.__main__:run_server"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]